# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Minecraft 皮肤/头像/服务器状态反向代理服务",
        "title": "MC Skin Wrap API",
        "contact": {
            "name": "VincentZyu"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "host": "127.0.0.1:60311",
    "basePath": "/gin_skin_wrap",
    "paths": {
        "/mcjava/avatar/{name}": {
            "get": {
                "description": "代理访问 https://minotar.net/avatar/{name}",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Minecraft"
                ],
                "summary": "获取玩家头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "头像图片",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/mcjava/server_status/{addr}": {
            "get": {
                "description": "代理访问 https://api.mcstatus.io/v2/status/java/{addr}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Minecraft"
                ],
                "summary": "获取服务器状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器地址",
                        "name": "addr",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器状态 JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mcjava/skin/{name}": {
            "get": {
                "description": "代理访问 https://minotar.net/skin/{name}",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Minecraft"
                ],
                "summary": "获取玩家皮肤",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "皮肤图片",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    }
}